+++ pennylane/math/__init__.py
@@ -33,14 +33,13 @@ The following frameworks are currently supported:
 """
 import autoray as ar
 
-from .classical import frobenius_inner_product
-
 from .multi_dispatch import (
     _multi_dispatch,
     block_diag,
     concatenate,
     diag,
     dot,
+    frobenius_inner_product,
     ones_like,
     stack,
     where,
